+++ examples/client.py
@@ -3,7 +3,7 @@ import asyncio
 import logging
 import random
 
-from cabbage import AmqpConnection, AsyncAmqpRpc
+from cabbagok import AmqpConnection, AsyncAmqpRpc
 
 logging.basicConfig(
     level=logging.DEBUG,