+++ tests/test_server.py
@@ -1,13 +1,14 @@
 # -*- coding: utf-8 -*-
+from unittest.mock import patch, MagicMock
+
 import aioamqp
 import asyncio
 import random
 from itertools import islice
 
 import pytest
-from asynctest import patch, MagicMock
 
-import cabbage
+import cabbagok
 from tests.conftest import MockTransport, MockProtocol, SUBSCRIPTION_QUEUE, TEST_EXCHANGE, SUBSCRIPTION_KEY, \
     RANDOM_QUEUE, HOST, MockEnvelope, MockProperties, CONSUMER_TAG, DELIVERY_TAG, RESPONSE_CORR_ID
 
@@ -20,10 +21,10 @@ class TestConnect:
 
     @pytest.mark.parametrize('exchange', ['', 'seldom-exchange', 'public', 'private'])
     async def test_ok(self, event_loop, exchange):
-        connection = cabbage.AmqpConnection(hosts=[(HOST, 5672)], loop=event_loop)
-        rpc = cabbage.AsyncAmqpRpc(connection=connection)
+        connection = cabbagok.AmqpConnection(hosts=[(HOST, 5672)], loop=event_loop)
+        rpc = cabbagok.AsyncAmqpRpc(connection=connection)
         rpc.callback_exchange = exchange
-        with patch('cabbage.amqp.aioamqp_connect') as mock_connect:
+        with patch('cabbagok.amqp.aioamqp_connect') as mock_connect:
             mock_connect.return_value = (MockTransport(), MockProtocol())
             await rpc.connect()
         mock_connect.assert_called_once()
@@ -50,7 +51,7 @@ class TestSubscribe:
         def request_handler(request):
             return request
 
-        rpc = cabbage.AsyncAmqpRpc(
+        rpc = cabbagok.AsyncAmqpRpc(
             connection=connection,
             queue_params=dict(passive=False, durable=True, exclusive=True, auto_delete=True),
             exchange_params=dict(type_name='fanout', passive=False, durable=True, auto_delete=True))
@@ -74,7 +75,7 @@ class TestSubscribe:
         #     callback=partial(rpc._on_request, request_handler=request_handler), queue_name=SUBSCRIPTION_QUEUE)
 
     async def test_defaults(self, connection):
-        rpc = cabbage.AsyncAmqpRpc(connection=connection)
+        rpc = cabbagok.AsyncAmqpRpc(connection=connection)
         await rpc.connect()
         assert rpc.channel.queue_declare.call_count == 1
         assert rpc.channel.basic_consume.call_count == 1
@@ -95,7 +96,7 @@ class TestSubscribe:
 
     async def test_amqp_defaults(self, connection):
         """Test that broker defaults (empty queue, empty exchange) are handled well."""
-        rpc = cabbage.AsyncAmqpRpc(connection=connection)
+        rpc = cabbagok.AsyncAmqpRpc(connection=connection)
         await rpc.connect()
         assert rpc.channel.queue_declare.call_count == 1
         assert rpc.channel.basic_consume.call_count == 1
@@ -119,7 +120,7 @@ class TestUnsubscribe:
     """AsyncAmqpRpc.unsubscribe"""
 
     async def test_ok(self, connection):
-        rpc = cabbage.AsyncAmqpRpc(connection=connection)
+        rpc = cabbagok.AsyncAmqpRpc(connection=connection)
         await rpc.connect()
         await rpc.unsubscribe(consumer_tag=CONSUMER_TAG)
         rpc.channel.basic_cancel.assert_called_once_with(consumer_tag=CONSUMER_TAG)
@@ -148,7 +149,7 @@ class TestHandleRpc:
     ])
     async def test_responding(self, connection, body, expected, is_async):
         handler = self.request_handler_factory(is_async, responding=True, fail=False)
-        rpc = cabbage.AsyncAmqpRpc(connection=connection)
+        rpc = cabbagok.AsyncAmqpRpc(connection=connection)
         await rpc.connect()
         await rpc.handle_rpc(channel=rpc.channel, body=body, envelope=MockEnvelope(), properties=MockProperties(),
                              request_handler=handler)
@@ -166,7 +167,7 @@ class TestHandleRpc:
     async def test_not_responding(self, connection, body, expected, is_async):
         """Handler returns None instead of str/bytes => no response needed."""
         handler = self.request_handler_factory(is_async, responding=False, fail=False)
-        rpc = cabbage.AsyncAmqpRpc(connection=connection)
+        rpc = cabbagok.AsyncAmqpRpc(connection=connection)
         await rpc.connect()
         await rpc.handle_rpc(channel=rpc.channel, body=body, envelope=MockEnvelope(), properties=MockProperties(),
                              request_handler=handler)
@@ -182,7 +183,7 @@ class TestHandleRpc:
     ])
     async def test_exception(self, connection, body, expected, is_async):
         handler = self.request_handler_factory(is_async, fail=True)
-        rpc = cabbage.AsyncAmqpRpc(connection=connection)
+        rpc = cabbagok.AsyncAmqpRpc(connection=connection)
         await rpc.connect()
         await rpc.handle_rpc(channel=rpc.channel, body=body, envelope=MockEnvelope(), properties=MockProperties(),
                              request_handler=handler)
@@ -194,7 +195,7 @@ class TestHandleRpc:
     @pytest.mark.parametrize('channel', [True, False])
     async def test_wait_connection(self, is_connected, channel):
         """
-        Checking for execution time of cabbage.AsyncAmqpRpc.wait_connected()
+        Checking for execution time of cabbagok.AsyncAmqpRpc.wait_connected()
         In case of True value of both variables 'connection.is_connected' and 'channel'
         the function should terminate immediately. Else the function should check
         the state of these variables with an interval determined in variable 'connection_delay' (in seconds)
@@ -212,7 +213,7 @@ class TestHandleRpc:
                 self.connection_delay = TEST_DELAY
 
         fake_self = FakeSelf(is_connected, channel)
-        future = asyncio.ensure_future(cabbage.AsyncAmqpRpc.wait_connected(fake_self))
+        future = asyncio.ensure_future(cabbagok.AsyncAmqpRpc.wait_connected(fake_self))
         await asyncio.sleep(TEST_DELAY)
         if not future.done():
             fake_self.connection.is_connected = True
@@ -225,14 +226,14 @@ class TestHandleRpc:
     @pytest.mark.parametrize('pending', [True, False])
     async def test_launch_server(self, connection, number_of_tasks, pending):
         """
-        Test for cabbage.AsyncAmqpRpc.stop(). All tasks should execute asynchronously.
+        Test for cabbagok.AsyncAmqpRpc.stop(). All tasks should execute asynchronously.
         In process of tests it may be created a big task (pending variable) if compare with others ones.
         In this case the task should continue the executing after calling the target function
         """
 
         small_delay = TEST_DELAY * 0.5
         big_delay = TEST_DELAY * 3
-        rpc = cabbage.AsyncAmqpRpc(connection=connection)
+        rpc = cabbagok.AsyncAmqpRpc(connection=connection)
         rpc.shutdown_timeout = TEST_DELAY
         await rpc.connect()
         await rpc.subscribe(request_handler=lambda x: x, queue=SUBSCRIPTION_QUEUE)
@@ -254,8 +255,8 @@ class TestHandleRpc:
 
     async def test_run(self):
         """
-        Test for cabbage.AsyncAmqpRpc.run(). This function requires
-        termination of cabbage.AsyncAmqpRpc.wait_connected() function
+        Test for cabbagok.AsyncAmqpRpc.run(). This function requires
+        termination of cabbagok.AsyncAmqpRpc.wait_connected() function
         """
 
         class FakeRunner:
@@ -271,16 +272,16 @@ class TestHandleRpc:
         delay = TEST_DELAY
         delta = 0.1 * TEST_DELAY
         fake_runner = FakeRunner(delay)
-        future = asyncio.ensure_future(cabbage.AsyncAmqpRpc.run(fake_runner))
+        future = asyncio.ensure_future(cabbagok.AsyncAmqpRpc.run(fake_runner))
 
         await asyncio.sleep(delay + delta)
 
-        # wait_connected() have been terminated and cabbage.AsyncAmqpRpc.run() should also terminate
+        # wait_connected() have been terminated and cabbagok.AsyncAmqpRpc.run() should also terminate
         assert future.done()
 
     async def test_run_server(self):
         """
-        Test for cabbage.AsyncAmqpRpc.run_server. The fuction shouldn't terminate until
+        Test for cabbagok.AsyncAmqpRpc.run_server. The fuction shouldn't terminate until
         the variable self.keep_running is True.
         """
 
@@ -314,7 +315,7 @@ class TestHandleRpc:
         test_delay = TEST_DELAY
         delta = TEST_DELAY * 0.1
         fake_self = FakeSelf(test_delay)
-        future = asyncio.ensure_future(cabbage.AsyncAmqpRpc.run_server(fake_self))
+        future = asyncio.ensure_future(cabbagok.AsyncAmqpRpc.run_server(fake_self))
 
         # While fake_self.keep_runnig the function shouldn't terminate
         await asyncio.sleep(3 * test_delay)
@@ -329,7 +330,7 @@ class TestHandleRpc:
 
     async def test_on_request(self, rpc):
         """
-        Test for cabbage.AsyncAmqpRpc._on_request
+        Test for cabbagok.AsyncAmqpRpc._on_request
         It's checking that callback inside the function has been called
         """
         big_delay = 2 * TEST_DELAY
@@ -338,6 +339,7 @@ class TestHandleRpc:
             await asyncio.sleep(big_delay)
 
         delta = TEST_DELAY * 0.1
+        rpc = await rpc
         await rpc.connect()
         asyncio.ensure_future(rpc._on_request(rpc.channel, b'', MockEnvelope(), MockProperties(), run_delay))
 
@@ -352,7 +354,7 @@ class TestHandleRpc:
     @pytest.mark.parametrize('shuffle', [True, False])
     async def test_shuffle(self, shuffle):
         """
-        Test for cabbage.AmqpConnection.cycle_hosts
+        Test for cabbagok.AmqpConnection.cycle_hosts
         It checks correct work of parameter 'shuffle' and also changing the state of
         internal variable self.hosts during shuffling
         """
@@ -363,7 +365,7 @@ class TestHandleRpc:
 
         # Creating a copy, and use its in constructor of AmqpConnection
         hosts_copy = hosts.copy()
-        connection = cabbage.AmqpConnection(hosts_copy)
+        connection = cabbagok.AmqpConnection(hosts_copy)
 
         # Retrieving one period of hosts after cycle hosts
         retrieved_hosts_two_periods = list(islice(connection.cycle_hosts(shuffle), hosts_length * 2))